test(auto-reply): split the thinking directive tests below the line cap
